Create Chaotic Systems, Tent Maps
The behavior of fixed point iterations lead to some of the first examples encountered when studying chaotic systems.  Consider the tent map on the interval[0,1]:f(x) =a(0.5−|x−0.5|), with a∈(0,2].
(a) How many fixed points does f have on the interval [0,1]?  What are they (in terms of a)?
(b)  Are they stable? Unstable?
(c)  Write a small code to perform the following fixed point iteration. For a= 2, make a table of  the  first  10  iterates xk+1=f(xk) for two starting values: x0= 0.399 and x0= 0.400. 
What is happening?  If you run the fixed point iteration for 200 iterates, what happens and why?