Create Backward Substitution Implementation
Write  a  code  for  backward  substitution  to  solve  systems  of  the  form Ux=b,  i.e.,  write  a  function x = backward(A,b),  which  expects  as  inputs  an  upper triangular  matrix U∈R n×n,  and  a  right  hand  side  vector b∈Rn,  which  returns the solution vector x∈Rn.  The function should find the size n from the vector b and also check if the matrix and the vector sizes are compatible before it starts to solve the system.